refactor(frontend): move hooks into their own files
Cleans up FileViewPage.tsx by moving general-purpose hooks into their own files.
